split simulation
